requirements.txt
anthropic
google-api-python-client
google-auth
google-auth-httplib2
sendgrid
python-dotenv
